Refactor clauses in a different package. * Introduction of Scalog
