+++ core/src/main/scala/io/github/scalaquest/core/parsing/scalog/Clause.scala
@@ -1,4 +1,4 @@
-package io.github.scalaquest.core.parsing.engine
+package io.github.scalaquest.core.parsing.scalog
 
 /** A Prolog clause. */
 sealed trait Clause {
@@ -64,11 +64,32 @@ case class Fact(body: Term) extends Clause {
    */
   def -->(right: Term): DCGRule = DCGRule(body, right)
 
+  /**
+   * Allows the creation of a [[Rule]].
+   * @param right the right side of the operation
+   * @return the resulting rule.
+   */
   def :-(right: Term): Rule = Rule(body, right)
 
   override def generate: String = s"${body.generate}."
 }
 
+/**
+ * A Prolog rule.
+ *
+ * @param head the left term
+ * @param body the right term
+ *
+ * Example:
+ * {{{
+ *   Rule(Compound(Atom("hello"), Atom("world")), Atom("hello")).generate
+ * }}}
+ *
+ * Should result in the following Prolog clause being printed:
+ * {{{
+ *   hello(world) :- hello.
+ * }}}
+ */
 case class Rule(head: Term, body: Term) extends Clause {
   override def generate: String = s"${head.generate} :- ${body.generate}."
 }
@@ -80,7 +101,7 @@ case class Rule(head: Term, body: Term) extends Clause {
  *
  * Example:
  * {{{
- *   DCGRule(Compound(Atom("ciao"), Atom("mondo")), ListP(Atom("ciao"), Atom("mondo"))).generate
+ *   DCGRule(Compound(Atom("hello"), Atom("world")), ListP(Atom("hello"), Atom("world"))).generate
  * }}}
  *
  * Should result in the following Prolog clause being printed:
@@ -90,68 +111,4 @@ case class Rule(head: Term, body: Term) extends Clause {
  */
 case class DCGRule(left: Term, right: Term) extends Clause {
   override def generate: String = s"${left.generate} --> ${right.generate}."
-}
-
-/** This object is mimicking a package declaration. */
-object clause {
-
-  /**
-   * This package includes some facility methods in order to create terms in a
-   * Prolog-like syntax.
-   *
-   * Example:
-   * {{{
-   *   val hello = CompoundBuilder("hello")
-   *   val X = Variable("X")
-   *   (hello(X) --> ListP("hello", X)).generate
-   * }}}
-   *
-   * Should result in the following Prolog clause being printed:
-   * {{{
-   *   hello(X) --> ["hello", X].
-   * }}}
-   */
-  object dsl {
-
-    /**
-     * A `CompoundBuilder` allows the generation of compound terms using a
-     * Prolog-like syntax that is checked by the scala compiler.
-     * @param functor the atom used as a functor by the generated compound terms
-     *
-     *  Example:
-     *  {{{
-     *    val hello = CompoundTerm(Atom("hello"))
-     *    (hello(Atom("darkness"), Atom("my"), Atom("old"), Atom("friend"))).generate
-     *  }}}
-     *
-     *  Should result in the following compound term being printed:
-     *  {{{
-     *    hello(darkness, my, old, friend)
-     *  }}}
-     */
-    case class CompoundBuilder(functor: Atom) {
-
-      /**
-       * Allows to simulate the Prolog compound term declaration.
-       * @param arg the first mandatory term
-       * @param args the other optional terms
-       * @return a [[Compound]] which has `functor` as functor and the arguments
-       *         as the compound term arguments.
-       */
-      def apply(arg: Term, args: Term*): Compound = Compound(functor, arg, args.toList)
-    }
-
-    /** Enables the implicit conversion from `String` to `Atom` */
-    implicit def stringToAtom(name: String): Atom = Atom(name)
-
-    /** Enables the implicit conversion from `Int` to `Number` */
-    implicit def intToNumber(n: Int): Number = Number(n)
-
-    /** Enables the implicit conversion from `Term` to `Fact` */
-    implicit def termToFact(term: Term): Fact = Fact(term)
-
-    /** Enables the implicit conversion fron `Seq[Term]` to `ListP` */
-    implicit def seqToListP(seq: Seq[Term]): ListP = ListP(seq: _*)
-
-  }
 }
