+++ core/src/test/scala/io/github/scalaquest/core/parsing/scalog/ClausesTest.scala
@@ -1,4 +1,4 @@
-package io.github.scalaquest.core.parsing.engine
+package io.github.scalaquest.core.parsing.scalog
 
 import org.scalatest.wordspec.AnyWordSpec
 
@@ -95,7 +95,7 @@ class ClausesTest extends AnyWordSpec {
   }
 
   "Compound builder" should {
-    import io.github.scalaquest.core.parsing.engine.clause.dsl.CompoundBuilder
+    import io.github.scalaquest.core.parsing.scalog.dsl.CompoundBuilder
     val hello = CompoundBuilder(Atom("hello"))
 
     "allow for the creation of a compound term" in {
@@ -115,7 +115,7 @@ class ClausesTest extends AnyWordSpec {
   }
 
   "Implicit operators" should {
-    import io.github.scalaquest.core.parsing.engine.clause.dsl._
+    import io.github.scalaquest.core.parsing.scalog.dsl._
     val hello = CompoundBuilder(Atom("hello"))
     "allow the usage of strings as atoms" in {
       assert(hello("world") == Compound(Atom("hello"), Atom("world")))
